refactor and put controls in their own files
