+++ stories/features/shields2/components/privacyControls.tsx
@@ -4,26 +4,10 @@
 
 import * as React from 'react'
 
-// Feature-specific components
-import {
-  BlockedInfoRow,
-  BlockedInfoRowForSelect,
-  BlockedInfoRowSingle,
-  BlockedInfoRowData,
-  BlockedInfoRowDataForSelect,
-  ArrowDownIcon,
-  BlockedInfoRowStats,
-  BlockedInfoRowText
-} from '../../../../src/features/shields2'
-import { Toggle, SelectBox } from '../../../../src/features/shields'
-
 // Group Components
-import StaticList from './list/static'
-import DynamicList from './list/dynamic'
-
-// Fake data
-import { getLocale } from '../fakeLocale'
-import data from '../fakeData'
+import ScriptsControl from './controls/scriptsControl'
+import CookiesControl from './controls/cookiesControl'
+import DeviceRecognitionControl from './controls/deviceRecognitionControl'
 
 interface Props {
   favicon: string
@@ -41,128 +25,34 @@ interface State {
 }
 
 export default class PrivacyControls extends React.PureComponent<Props, State> {
-  constructor (props: Props) {
-    super(props)
-    this.state = {
-      deviceRecognitionOpen: false,
-      scriptsBlockedOpen: false,
-      scriptsBlockedEnabled: true
-    }
-  }
-
-  get tabIndex () {
-    const { isBlockedListOpen } = this.props
-    return isBlockedListOpen ? -1 : 0
-  }
-
-  onOpenScriptsBlockedOpen = (event: React.MouseEvent<HTMLDivElement>) => {
-    if (event.currentTarget) {
-      event.currentTarget.blur()
-    }
-    this.props.setBlockedListOpen()
-    this.setState({ scriptsBlockedOpen: !this.state.scriptsBlockedOpen })
-  }
-
-  onOpenScriptsBlockedOpenViaKeyboard = (event: React.KeyboardEvent<HTMLDivElement>) => {
-    if (event) {
-      if (event.key === ' ') {
-        event.currentTarget.blur()
-        this.props.setBlockedListOpen()
-        this.setState({ scriptsBlockedOpen: !this.state.scriptsBlockedOpen })
-      }
-    }
-  }
-
-  onChangeScriptsBlockedEnabled = (event: React.ChangeEvent<HTMLInputElement>) => {
-    this.setState({ scriptsBlockedEnabled: event.target.checked })
-  }
-
-  onOpenDeviceRecognitionOpen = (event: React.MouseEvent<HTMLDivElement>) => {
-    if (event.currentTarget) {
-      event.currentTarget.blur()
-    }
-    this.props.setBlockedListOpen()
-    this.setState({ deviceRecognitionOpen: !this.state.deviceRecognitionOpen })
-  }
-
-  onOpenDeviceRecognitionOpenViaKeyboard = (event: React.KeyboardEvent<HTMLDivElement>) => {
-    if (event) {
-      if (event.key === ' ') {
-        event.currentTarget.blur()
-        this.props.setBlockedListOpen()
-        this.setState({ deviceRecognitionOpen: !this.state.deviceRecognitionOpen })
-      }
-    }
-  }
-
   render () {
-    const { favicon, hostname, isBlockedListOpen, scriptsBlocked, fingerprintingBlocked } = this.props
     const {
-      deviceRecognitionOpen,
-      scriptsBlockedEnabled,
-      scriptsBlockedOpen
-    } = this.state
+      favicon,
+      hostname,
+      setBlockedListOpen,
+      isBlockedListOpen,
+      scriptsBlocked,
+      fingerprintingBlocked
+    } = this.props
     return (
       <>
-        <BlockedInfoRow>
-          <BlockedInfoRowData
-            tabIndex={this.tabIndex}
-            onClick={this.onOpenScriptsBlockedOpen}
-            onKeyDown={this.onOpenScriptsBlockedOpenViaKeyboard}
-          >
-            <ArrowDownIcon />
-            <BlockedInfoRowStats>{scriptsBlocked > 99 ? '99+' : scriptsBlocked}</BlockedInfoRowStats>
-            <BlockedInfoRowText>{getLocale('scriptsBlocked')}</BlockedInfoRowText>
-          </BlockedInfoRowData>
-          <Toggle
-            disabled={isBlockedListOpen}
-            checked={scriptsBlockedEnabled}
-            onChange={this.onChangeScriptsBlockedEnabled}
-          />
-          {
-            scriptsBlockedOpen &&
-              <DynamicList
-                favicon={favicon}
-                hostname={hostname}
-                name={getLocale('scriptsOnThisSite')}
-                list={data.blockedScriptsResouces}
-                onClose={this.onOpenScriptsBlockedOpen}
-              />
-          }
-        </BlockedInfoRow>
-        <BlockedInfoRowSingle>
-            <SelectBox disabled={isBlockedListOpen}>
-              <option value='block_third_party'>{getLocale('thirdPartyCookiesBlocked')}</option>
-              <option value='block'>{getLocale('allCookiesBlocked')}</option>
-              <option value='allow'>{getLocale('allCookiesAllowed')}</option>
-            </SelectBox>
-        </BlockedInfoRowSingle>
-        <BlockedInfoRowForSelect>
-          <BlockedInfoRowDataForSelect
-            tabIndex={this.tabIndex}
-            onClick={this.onOpenDeviceRecognitionOpen}
-            onKeyDown={this.onOpenDeviceRecognitionOpenViaKeyboard}
-          >
-            <ArrowDownIcon />
-            <BlockedInfoRowStats>{fingerprintingBlocked > 99 ? '99+' : fingerprintingBlocked}</BlockedInfoRowStats>
-          </BlockedInfoRowDataForSelect>
-          <SelectBox disabled={isBlockedListOpen}>
-            <option value='block_third_party'>{getLocale('thirdPartyFingerprintingBlocked')}</option>
-            <option value='block'>{getLocale('allFingerprintingBlocked')}</option>
-            <option value='allow'>{getLocale('allFingerprintingAllowed')}</option>
-          </SelectBox>
-          {
-            deviceRecognitionOpen &&
-              <StaticList
-                favicon={favicon}
-                hostname={hostname}
-                stats={fingerprintingBlocked}
-                name={getLocale('deviceRecognitionAttempts')}
-                list={data.blockedFakeResources}
-                onClose={this.onOpenDeviceRecognitionOpen}
-              />
-          }
-        </BlockedInfoRowForSelect>
+        <ScriptsControl
+          favicon={favicon}
+          hostname={hostname}
+          isBlockedListOpen={isBlockedListOpen}
+          scriptsBlocked={scriptsBlocked}
+          setBlockedListOpen={setBlockedListOpen}
+        />
+        <CookiesControl
+          isBlockedListOpen={isBlockedListOpen}
+        />
+        <DeviceRecognitionControl
+          favicon={favicon}
+          hostname={hostname}
+          isBlockedListOpen={isBlockedListOpen}
+          fingerprintingBlocked={fingerprintingBlocked}
+          setBlockedListOpen={setBlockedListOpen}
+        />
       </>
     )
   }
